Todo App done with Docker(for backend: Django only)

+++ ReactJs-Django-GraphQL Projects/TODO_APP/backend/todo/schema.py
@@ -1,22 +1,35 @@
-from django.core.checks import messages
 import graphene
-from graphene_django.fields import DjangoListField
 from .models import Todo
 from graphene_django import DjangoObjectType
+from graphql_jwt.decorators import login_required
+from django.contrib.auth import get_user_model
 
 class TodoType(DjangoObjectType):
     class Meta:
         model=Todo
         fields=('id','title','date')
 
+class UserType(DjangoObjectType):
+    class Meta:
+        model=get_user_model()
+
+
 
 class Query(graphene.ObjectType):
     todos=graphene.List(TodoType,id=graphene.Int())
-
+    user = graphene.Field(UserType)
+    @login_required
     def resolve_todos(self,info,id=None):
+        user=info.context.user
         if id:
             return Todo.objects.filter(id=id)
-        return Todo.objects.all().order_by("-id")
+        return Todo.objects.filter(user=user).order_by("-id")
+    @login_required
+    def resolve_user(self, info):
+        user = info.context.user
+        if user.is_anonymous:
+            raise Exception("Login Required")
+        return user
 
 class CreateTodo(graphene.Mutation):
     todo = graphene.Field(TodoType)
@@ -24,8 +37,12 @@ class CreateTodo(graphene.Mutation):
     class Arguments:
         title = graphene.String(required=True)
 
+    @login_required
     def mutate(self, info, title):
-        todo = Todo(title=title)
+        user=info.context.user
+        if user.is_anonymous:
+            raise Exception("Not logged in ! Login Now")
+        todo = Todo(user=user,title=title)
         todo.save()
         return CreateTodo(todo=todo)
 
@@ -37,8 +54,12 @@ class UpdateTodo(graphene.Mutation):
         id=graphene.Int(required=True)
         title = graphene.String(required=True)
 
+    @login_required
     def mutate(self, info,id, title):
+        user=info.context.user
         todo=Todo.objects.get(id=id)
+        if user!=todo.user:
+            raise Exception("It's not your todo item !")
         todo.title = title
         todo.save()
         return UpdateTodo(todo=todo)
@@ -49,13 +70,29 @@ class DeleteTodo(graphene.Mutation):
     class Arguments:
         id=graphene.Int(required=True)
 
+    @login_required
     def mutate(self,info,id):
+        user=info.context.user
         todo=Todo.objects.get(id=id)
+        if user!=todo.user:
+            raise Exception("It's not your todo item !")
         todo.delete()
         return DeleteTodo(message=f"ID {id} deleted")
+
+class CreateUser(graphene.Mutation):
+    user=graphene.Field(UserType)
+    class Arguments:
+        username=graphene.String(required=True)
+        password=graphene.String(required=True)
+    def mutate(self,info,username,password):
+        user=get_user_model()(username=username)
+        user.set_password(password)
+        user.save()
+        return CreateUser(user=user)
 
 
 class Mutation(graphene.ObjectType):
     create_todo=CreateTodo.Field()
     update_todo=UpdateTodo.Field()
     delete_todo=DeleteTodo.Field()
+    create_user=CreateUser.Field()
